refactor: update GoDAM logos and block icons to new Figma design (#1989)
* refactor: update GoDAM logos and block icons to new Figma design

- Replace all 4 Gutenberg block icons (video, gallery, audio, pdf) with new
  flat purple (#5D31FF) background + white icon design from Figma
- Copy updated block icons to assets/images/ (Elementor + WP Bakery)
- Add godam-logo.svg with new full logo (icon + GoDAM wordmark) in dark navy
- Update godam-icon.svg, godam-icon-16.svg, godam-logo-gradient.svg to new icon shape
- Update admin menu inline SVG (class-pages.php) base64 to new icon
- Update PHP admin pages to reference godam-logo.svg instead of godam-logo.png
- Update React page imports to reference godam-logo.svg / godam-icon.svg
- Rebuild Gutenberg blocks (npm run build:blocks)

* Add logo and fix alignment

* Add color to badges

* Add color to icons in media dn video editor
