+++ assets/src/css/admin.scss
@@ -907,7 +907,7 @@ to ensure proper responsiveness and avoid layout issues.
 	text-transform: uppercase;
 	letter-spacing: 0.05em;
 	color: #fff;
-	background: linear-gradient(83.85deg, #AB3A6C -9.3%, #E6533A 120.31%);
+	background: #5D31FF;
 	border-radius: 0.25rem;
 	vertical-align: middle;
 	line-height: 1.5;
