+++ inc/classes/class-pages.php
@@ -168,7 +168,7 @@ class Pages {
 			'edit_pages',
 			$this->menu_slug,
 			array( $this, 'render_dashboard_page' ),
-			'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNjQiIGhlaWdodD0iNjQiIHZpZXdCb3g9IjAgMCA2NCA2NCIgZmlsbD0ibm9uZSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj4KPHBhdGggZD0iTTI1LjU1NzggMjAuMDkxMUw4LjA1NTg3IDM3LjU5M0wzLjQ2Mzk3IDMzLjAwMTFDMC44MTg1MjEgMzAuMzU1NiAyLjA4MjEgMjUuODMzNiA1LjcyMjI4IDI0Ljk0NjRMMjUuNTYzMiAyMC4wOTY0TDI1LjU1NzggMjAuMDkxMVoiIGZpbGw9IndoaXRlIi8+CjxwYXRoIGQ9Ik00Ny4zNzczIDIxLjg4NjdMNDUuNTQzOCAyOS4zODc1TDIyLjY5NzIgNTIuMjM0MUwxMS4yNjA1IDQwLjc5NzRMMzQuMTY2MiAxNy44OTE2TDQxLjU3MDMgMTYuMDc5NkM0NS4wNzA2IDE1LjIyNDcgNDguMjMyMyAxOC4zODYzIDQ3LjM3MiAyMS44ODEzTDQ3LjM3NzMgMjEuODg2N1oiIGZpbGw9IndoaXRlIi8+CjxwYXRoIGQ9Ik00My41MDU5IDM4LjEwMzZMMzguNjY2NyA1Ny44OTA3QzM3Ljc3NDEgNjEuNTI1NSAzMy4yNTIxIDYyLjc4OTEgMzAuNjA2NiA2MC4xNDM2TDI2LjAzNjMgNTUuNTczMkw0My41MDU5IDM4LjEwMzZaIiBmaWxsPSJ3aGl0ZSIvPgo8L3N2Zz4K',
+			'data:image/svg+xml;base64,PHN2ZyB2aWV3Qm94PSItNSAtNSAxMzggMTM4IiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBkPSJNNjYuODY2NyAxMy44ODY2TDE1LjMxODMgNjUuNDgyNkwzLjM1OTYyIDUzLjUxMjlDLTMuNTI5OTEgNDYuNjE3IDAuNzQ3Mjc3IDMzLjg0MjEgMTAuOTk5MyAzMC43NTY4TDY2Ljg4MDcgMTMuOTAwNkw2Ni44NjY3IDEzLjg4NjZaIiBmaWxsPSJ3aGl0ZSIvPgo8cGF0aCBkPSJNMTI3LjQ0NCAxNi4wMzI4TDEyMS4zMTMgMzcuMjYzNkw1NC4zOTA4IDEwNC4yNDdMMjMuOTU4MyA3My43ODY4TDkxLjA1MzUgNi42Mjk2NEwxMTEuOTkxIDAuNTY2MjY3QzEyMS44OSAtMi4yOTU0NSAxMzAuMzAzIDYuMTI1MjUgMTI3LjQyOSAxNi4wMTg1TDEyNy40NDQgMTYuMDMyOFoiIGZpbGw9IndoaXRlIi8+CjxwYXRoIGQ9Ik0xMTQuMDEzIDYxLjA3NjlMOTcuMjA0NyAxMTYuOTg2Qzk0LjEwNjkgMTI3LjI1NyA4MS4zMTUgMTMxLjU0IDc0LjQwNDIgMTI0LjYyM0w2Mi40NjQ4IDExMi42NzNMMTE0LjAxMyA2MS4wNzY5WiIgZmlsbD0id2hpdGUiLz4KPC9zdmc+',
 			11
 		);
