+++ pages/video-editor/components/video-dataview/video-dataview.scss
@@ -338,9 +338,6 @@ $godam-border: rgba(246, 246, 246, 1);
 			min-width: 1.25rem;
 			min-height: 1.25rem;
 			object-fit: contain;
-			position: relative;
-			bottom: 1px;
-			left: 1px;
 		}
 	}
